Relocate TodoItem class
