+++ app/src/main/java/dev/msfjarvis/todo/data/model/TodoItem.kt
@@ -1,3 +1,3 @@
-package dev.msfjarvis.todo.data
+package dev.msfjarvis.todo.data.model
 
 data class TodoItem(val title: String)
